Creado get de usuarios y comprobado que se pueden añadir usuarios para implementar la validación
