--- Servidor/routes/usersRoutes.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { register, login } = require('../controllers/usersController');
-
-router.post('/', createUser);
-
-
-module.exports = router;
