+++ src/css/variables.css
@@ -11,9 +11,6 @@
     --s: 14px;
     --xs: 12px;
 
-    /* Tipografias */
-
-
     /* Espacios */
     --espaciado-grande: 1.5em;
     --espaciado-normal: 1em;
